Create tests for GameBoard instances
I created two tests: 1. Execute hits on some ship blocks of the
GameBoard instance and check that the game is not over 2. Execute hits
on all ship blocks of the GameBoard instance and check that the game is
indeed over. Additionally, I removed many of the comments for these two
tests and kept them for myself for use in the README. Finally, similar
to the Ship instance tests, I created a gameBoardFixture function that
creates a gameBoard instance and executes an amount of hits based on
the requirements of the test.