Start creating Player factory function
Fixed gameBoardFixture() in test and GameBoard() factory function in
index.js to handle only the arguments that are passed through so that
all tests don't have to create mock boards will all ship types. I also
made a Player factory function prototype that returns the value of the
chosen coordinate to attack. The first test for the Player factory
function passes.