Add attack and receiveAttackInfo methods to AI
The AI attack method still needs to be completed. However, the
"AI prioritizes attacking blocks adjacent to successful hits" test now
passes, because that part of the AI attack method was written along side
a functioning receiveAttackInfo method. The AI tests had to be slightly
modified however, because the or statements paired with the expect
statements were not written properly.