Design page layout
I created a rough design for my game using an test html file and a css
file that will be kept and used for this project. The next step will
be to create tests and then create the main game loop and the dom
manipulation. The html used will be converted to JavaScript when working
on the dom manipulation