Update gameboard instance receiveAttack() method
This method previously only returned strings "hit", or "miss". Now, the
method returns an array of two values: the attack status ("hit", "miss",
"sink"), and the type of ship hit. I did this because now that I am
developing the main game loop and dom manipulation functions, I realized
that my program needs to communicate what type of ship got hit in order
to communicate that info to the user.