Add configuration for webpack to handle images
I added configuration in the webpack.config.js file so that it can
handle images. However, the test for the missMarker() function still
does not pass, and I also tried to use the function in my index.js file,
and it still doesn't work. In the mean time, I created the
removeShipLifeBlock() function, and that passes its test. I'm not sure
if jest and webpack are properly handling these images, so this is an
issue I need to figure out.