Abandon dom tests
I realized that trying to test the dom functions was proving to be
difficult. I was forced to use the getComputedStyles() method, because
original styles were set through the style.css file, which was
miss matching the function mocks and the test. Because the mocks weren't
working, the only other option would be to use another jest test
environment, which I think is outside the scope of this project. Thi is
why I  decided to drop the dom-manipulation tests. I was also having
issues with implementing the images in my webpage, so I decided to drop
those and just use colors instead.