Create randomize function for ship placement
I decided against adding an option for the player to chose ship
placements through either drag-and-drop or entering coordinates. I did
this because the main point of this project was to practice Test Driven
Development, and in order to save time, I chose the randomizing of ship
positions. The implementation of choosing ship positions can be a
future, potential improvement, and will be discussed in the README.